export and builtins done

+++ libs/libft/ft_putstr_fd.c
@@ -6,7 +6,7 @@
 /*   By: eviscont <eviscont@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/28 18:18:55 by eviscont          #+#    #+#             */
-/*   Updated: 2023/10/10 16:22:13 by eviscont         ###   ########.fr       */
+/*   Updated: 2024/08/22 20:18:21 by eviscont         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putstr_fd(char *s, int fd)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	if (s == 0)
