+++ libs/libft/ft_substr.c
@@ -6,16 +6,16 @@
 /*   By: eviscont <eviscont@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/25 12:59:48 by eviscont          #+#    #+#             */
-/*   Updated: 2024/08/07 17:59:04 by eviscont         ###   ########.fr       */
+/*   Updated: 2024/08/22 20:07:05 by eviscont         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_substr(char const *s, unsigned int start, size_t len)
+char	*ft_substr(char const *s, int start, int len)
 {
-	char			*sb;
-	unsigned int	i;
+	char	*sb;
+	int		i;
 
 	i = 0;
 	if (s == 0)
